+++ src/Import/MsfsPln.ts
@@ -41,6 +41,7 @@ export class MsfsPln extends GarminFpl {
         lat: coords.lat,
         lon: coords.lon,
         alt: coords.altitude_ft,
+        countryCode: this.getXmlNode(xml, "ICAORegion") || undefined,
       };
     });
   }
@@ -116,8 +117,6 @@ export class MsfsPlnExport {
 
   toString(): string {
     const m = this.mission;
-    // @see https://en.wikipedia.org/wiki/ICAO_airport_code#/media/File:ICAO_FirstLetter.svg
-    let icaoRegion = m.origin_icao.substring(0, 1);
 
     const departureRunwayCp = m.findCheckPointByType(MissionCheckpoint.TYPE_DEPARTURE_RUNWAY);
     const departureRunway = departureRunwayCp ? departureRunwayCp.name : "";
@@ -159,13 +158,6 @@ export class MsfsPlnExport {
       ) {
         name = "RW" + name;
       }
-      if (
-        type === "Airport" ||
-        cp.type === MissionCheckpoint.TYPE_ORIGIN ||
-        cp.type === MissionCheckpoint.TYPE_DESTINATION
-      ) {
-        icaoRegion = cp.name.substring(0, 1);
-      }
 
       pln += `\
         <ATCWaypoint id="${Quote.xml(name)}">
@@ -182,7 +174,7 @@ export class MsfsPlnExport {
         pln += `\
             <ICAO>
                 <ICAOIdent>${Quote.xml(cp.name)}</ICAOIdent>
-                <ICAORegion>${Quote.xml(icaoRegion)}</ICAORegion>
+                <ICAORegion>${Quote.xml(cp.icao_region || "")}</ICAORegion>
             </ICAO>
 `;
       }